backend/requirements.txt
langchain
langchain-core
langchain_community
langchain_huggingface
langchain_text_splitters
langchain-google-genai
fastapi
python-dotenv
uvicorn
numpy
pandas
matplotlib
faiss-cpu
sentence-transformers
pypdf
routes
